fix(health): follow repository path semantics
